Add indicator to the generation(diffusion) model
DM model needs to know the condition type of the input data.
This commit adds an indicator to the generation model to indicate the condition type of the condition sequences.
The indicator is a learnable token with the length of the number of condition types.
The indicator is added with the input data before feeding it to the model.